refactored layout component and improved storybook
