+++ packages/libs/react-ui/src/components/Layout/Box/Box.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { describe, expect, test } from 'vitest';
 
-import { Box } from '@components/Box';
+import { Box } from '@components/Layout/Box';
 
 describe('Box', () => {
   test('renders a box', () => {
