+++ packages/libs/react-ui/src/components/Layout/Stack/stories.css.ts
@@ -1,0 +1,11 @@
+import { vars } from '@theme/vars.css';
+import { styleVariants } from '@vanilla-extract/css';
+
+export const itemSizeClass = styleVariants(vars.sizes, (size) => {
+  return [
+    {
+      width: size,
+      height: size,
+    },
+  ];
+});
