+++ packages/libs/react-ui/src/components/Pagination/Pagination.stories.tsx
@@ -1,7 +1,7 @@
 import type { SystemIcon } from '@components/Icon';
+import { Stack } from '@components/Layout/Stack';
 import type { IPaginationProps } from '@components/Pagination';
 import { Pagination } from '@components/Pagination';
-import { Stack } from '@components/Stack';
 import type { Meta, StoryObj } from '@storybook/react';
 import React from 'react';
