+++ packages/libs/react-ui/src/components/index.ts
@@ -3,7 +3,6 @@ export type {
   IAccordionRootProps,
   IAccordionSectionProps,
 } from './Accordion';
-export type { IBoxProps } from './Box';
 export type { IBreadcrumbItemProps, IBreadcrumbsProps } from './Breadcrumbs';
 export type { IButtonProps } from './Button';
 export type { ICardProps } from './Card';
@@ -18,9 +17,14 @@ export type {
   ITextareaProps,
   InputWrapperStatus,
 } from './Form';
-export type { IGridContainerProps, IGridItemProps } from './Grid';
 export type { IIconProps } from './Icon';
 export type { IIconButtonProps } from './IconButton';
+export type {
+  IBoxProps,
+  IGridItemProps,
+  IGridRootProps,
+  IStackProps,
+} from './Layout';
 export type { ILinkProps } from './Link';
 export type { MaskOptions } from './MaskedValue';
 export type { IMaskedValueProps } from './MaskedValue/MaskedValue';
@@ -50,7 +54,6 @@ export type {
   IProfileSummaryRootProps,
 } from './ProfileSummary';
 export type { IProgressBarProps } from './ProgressBar';
-export type { IStackProps } from './Stack';
 export type {
   ITBodyProps,
   ITHeadProps,
@@ -72,7 +75,6 @@ export type {
 } from './Typography';
 
 export { Accordion } from './Accordion';
-export { Box } from './Box';
 export { Breadcrumbs } from './Breadcrumbs';
 export { Button } from './Button';
 export { Card } from './Card';
@@ -87,9 +89,9 @@ export {
   TextField,
   Textarea,
 } from './Form';
-export { Grid } from './Grid';
 export { ProductIcon, SystemIcon } from './Icon';
 export { IconButton } from './IconButton';
+export { Box, Grid, Stack } from './Layout';
 export { Link } from './Link';
 export { MaskedValue } from './MaskedValue/MaskedValue';
 export { maskValue } from './MaskedValue/utils';
@@ -100,7 +102,6 @@ export { Notification } from './Notification';
 export { Pagination } from './Pagination';
 export { ProfileSummary } from './ProfileSummary';
 export { ProgressBar } from './ProgressBar';
-export { Stack } from './Stack';
 export { Table } from './Table';
 export { Tabs } from './Tabs';
 export { Tag } from './Tag';
