+++ packages/libs/react-ui/src/index.ts
@@ -13,8 +13,8 @@ export type {
   ICardProps,
   IContentHeaderProps,
   IGradientTextProps,
-  IGridContainerProps,
   IGridItemProps,
+  IGridRootProps,
   IHeadingProps,
   IIconButtonProps,
   IIconProps,
